// template/react-starter-kit/src/.generated/api/model/supportControllerGetAdminSupportTicketsParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { SortDirection } from './sortDirection';
import type { SupportControllerGetAdminSupportTicketsSortItem } from './supportControllerGetAdminSupportTicketsSortItem';
import type { SupportTicketPriority } from './supportTicketPriority';
import type { SupportTicketStatus } from './supportTicketStatus';

export type SupportControllerGetAdminSupportTicketsParams = {
sort?: SupportControllerGetAdminSupportTicketsSortItem[];
direction?: SortDirection[];
search?: string;
page?: number;
/**
 * @maximum 100
 */
limit?: number;
status?: SupportTicketStatus;
priority?: SupportTicketPriority;
category?: string;
};