// template/react-starter-kit/src/.generated/api/model/supportControllerGetAdminSupportTicketsSortItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type SupportControllerGetAdminSupportTicketsSortItem = typeof SupportControllerGetAdminSupportTicketsSortItem[keyof typeof SupportControllerGetAdminSupportTicketsSortItem];


export const SupportControllerGetAdminSupportTicketsSortItem = {
  category: 'category',
  title: 'title',
  status: 'status',
  priority: 'priority',
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
  id: 'id',
} as const;